arm: [MVE] Fix operands order in vbicq_f [PR122223]
The operands of the floating-point version of vbicq were swapped, this
patch fixes this.

For this backport the testcase needs an adjustment: the code is less
optimized than with gcc-15, so we still generate the 0.0f constant and
a vbic instruction.  We actually check that the 0.0f constant is in
the right vbic parameter.

gcc/ChangeLog:
	PR target/122223
	* config/arm/mve.md (@mve_vbicq_f<mode>): Fix operands order.

gcc/testsuite/ChangeLog:
	PR target/122223
	* gcc.target/arm/mve/intrinsics/pr122223.c: New test.

	(cherry picked from commits
	81e226440381cc3e033df7e58cc7793c9b4b4e25 and
	a52888dc71924afb6cd187b0e5f18d2be4c68a07)