aarch64: mingw: Make long double 64 bit
On windows targets, the size of long double is 64 bits. Therefore,
unlike aarch64-**-linux-gnu, where the size of long double is 128
bits, the size of long double for aarch64-w64-mingw32 has to be 64
bits.

This commit makes changes to gcc so that long double is 64 bits for
aarch64-w64-mingw32.

gcc/ChangeLog:

	* config/aarch64/aarch64-abi-ms.h
	(TARGET_LONG_DOUBLE_128): Set this to 0.
	* config/aarch64/aarch64.cc
	(aarch64_scalar_mode_supported_p): Make long double 64 bits.
	(aarch64_c_mode_for_floating_type): Return true for TFmode.
	* config/aarch64/aarch64.h
	(TARGET_LONG_DOUBLE_128): Set this to 1.

gcc/testsuite/ChangeLog:

	* gcc.target/aarch64/mingw/mingw.exp: New test.
	* gcc.target/aarch64/mingw/long_double_size.c: New test.

co-authored-by: Radek Barton <radek.barton@microsoft.com>
co-authored-by: Martin Vejbora <mvejbora@microsoft.com>