[PR target/119587] RISC-V: xtheadmemidx: Split slli.uw pattern
The combine pass can generate an index like (and:DI (mult:DI (reg:DI)
(const_int scale)) (const_int mask)) when XTheadMemIdx is available.
LRA may pull it out, and thus a splitter is needed when Zba is not
available.

A similar splitter were introduced when XTheadMemIdx support was added,
but removed in commit 31c3c5d.  The new splitter in this new patch is
based on the removed one.

	PR target/119587
gcc/ChangeLog:

	* config/riscv/thead.md (*th_memidx_operand): New splitter.

gcc/testsuite/ChangeLog:

	* gcc.target/riscv/xtheadmemidx-bug.c: New test.