a68: documentation
This commit adds a new section to the GCC Internals Manual and also
adds two new manuals.

ga68.texi is The GNU Algol 68 Compiler user manual.  It describes how
to use the compiler and all the GNU extensions implemented on top of
the Algol 68 programming language.

ga68-internals.texi is The GNU algol68 Compiler Internals manual.  It
describes the implementation of the front-end and it is of interest
primarily for developers.

Signed-off-by: Jose E. Marchesi <jemarch@gnu.org>

gcc/ChangeLog

	* algol68/ga68-internals.texi: New file.
	* algol68/ga68.texi: Likewise.