c2y: Allow unspecified arrays in generic association.
To allow unspecified arrays in generic association add a new
declaration context GENERIC_ASSOC for grokdeclarator and new
function grokgenassoc to be used by the parser.  The error
about unspecified array is moved from build_array_declarator
to grokdeclarator to be able to check for this.

gcc/c/ChangeLog:
	* c-decl.cc (build_array_declarator): Remove error.
	(grokgenassoc): New function.
	(grokdeclarator): Add error.
	* c-parser.cc (c_parser_generic_selection): Use grokgenassoc.
	* c-tree.h (grokgenassoc): Add prototype.

gcc/testsuite/ChangeLog:
	* gcc.dg/c2y-generic-6.c: New test.
	* gcc.dg/c2y-generic-7.c: New test.