Objective-C/C++: Enable Wignored-attributes.
r16-4373 altered headers so that Wignored-attributes was named in
a diagnostic push.  This causes several Objective-C++ tests to fail
since the atomicity.h header is included there.

Since Objective-C/C++ are intended to be supersets of the base
language, there is no specific reason to exclude this warning there.

gcc/c-family/ChangeLog:

	* c.opt: Enable Wignored-attributes for Objective-C and
	Objective-C++.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>