Remove max_count computation from ipa-inline
Ipa inline computes max_count which used to be applied later to compute badness
before it was converted to sreal.  Now it is only used in couple of places to see
if any IPA profile is presents at all.  This patch replaces this by more specific
flag has_nonzero_ipa_profile.

gcc/ChangeLog:

	* ipa-inline.cc (max_count): Remove.
	(has_nonzero_ipa_profile): New.
	(inline_small_functions): Update.
	(dump_inline_stats): Update.