Add reduc_sbool_{and,ior,xor}_scal optabs
The following adds named patterns for reducing of vector masks with
AND, IOR and XOR to be used by the vectorizer.  A slight complication
are targets using scalar integer modes as mask modes, as for those
the mode for low-precision masks is ambiguous.  For this reason the
optab follows what vec_pack_sbool_trunc does and passes an additional
CONST_INT operand indicating the number of lanes in the input mask.
Note this is done always when the vector mask mode is an integer mode
and never otherwise.

	* doc/md.texi (reduc_sbool_{and,ior,xor}_scal_<mode>): Document.
	* optabs.def (reduc_sbool_and_scal_optab,
	reduc_sbool_ior_scal_optab, reduc_sbool_xor_scal_optab): New.
	* internal-fn.def (REDUC_SBOOL_AND, REDUC_SBOOL_IOR,
	REDUC_SBOO_XOR): Likewise.
	* internal-fn.cc (reduc_sbool_direct): New initializer.
	(expand_reduc_sbool_optab_fn): New expander.
	(direct_reduc_sbool_optab_supported_p): New.