ada: Remove workaround for a freezing issue with address of a slice
Rewriting of an address of a slice, e.g. "X (Low .. High)'Address" into an
address of an indexed component, e.g. "X (Low)'Address", was only done as a
workaround for some freezing issue. Apparently this freezing issue is now
solved and the rewriting is causing problems for GNATprove. (This rewriting,
if needed, should be rather done in expansion and not resolution.)

gcc/ada/ChangeLog:

	* sem_attr.adb (Resolve_Attribute): Remove rewriting of a slice
	address expression.