docs: fmv:  Update Function multi-versioning documentation [PR c/122202]
This updates the FMV documentation to the current state of things, including
the addition of "target_version" based FMV.

Left as much of the x86 target based FMV documentation unchanged as
the behaviour change there should be unchanged. Though highlights some of
the differences between it and target_version FMV to try avoid confusion there.

	PR c/122202

gcc/ChangeLog:

	* doc/extend.texi (target function attribute): Update to describe FMV
	behaviour.
	(target_version function attribute): New section.
	(target_clones attribute): Update to descrbe new behaviour with
	target_version.
	(Function Multiversioning): Update to discuss both target_version and
	target based FMV.