ada: Reject nonconfirming size clauses on objects of floating-point types
The RM 13.1(7/5) subclause contains the following sentences: "If the size
of an object is greater than that of its subtype, the additional bits are
padding bits. For an elementary object, these padding bits are normally
read and updated along with the others."

GNAT implements it only for objects of integer and fixed-point types.
For objects of floating-point types, the padding bits are ignored (and
therefore not initialized) with a warning:

  warning: 32 bits of "F" unused [enabled by default]

Implementing it for objects of floating-point types does not seem worth the
hassle, so rejecting nonconfirming size clauses is probably the cleanest
approach (clauses giving the object a size lower than that of its subtype
are already rejected for objects of floating-point types).

gcc/ada/ChangeLog:

	* sem_ch13.adb (Analyze_Attribute_Definition_Clause)
	<Attribute_Size>: Tweak wording of existing error message for the
	size of all elementary objects.  Move error handling for aliased
	objects into the block for objects and give a similar error for
	(nonaliased) floating-point objects.