[PATCH] Makefile.tpl: remove an extra \; from find command
The extra \; parameter in the find command causes it to fail immediately and
not clean any config.cache:

$ find . -name config.cache -exec rm -f {} \; \;
find: paths must precede expression: `;'

This is benign in most cases but the binutils is also using this Makefile.tpl and
as the result its 'make distclean' can leave config.cache files around, which
fails subsequent attempts to configure and build it.

I have modified the Makefile.tpl and regenerated Makefile.in from it. For testing
I ran a config/make/make distclean loop.

	* Makefile.tpl (distclean): Remove extraenous semicolon.
	* Makefile.in: Rebuilt.